Allow to silence "error handling stats line" messages
"error handling stats line" messages can be really noisy and can be
ignored ignored in many cases (where module stats are not required/wanted).

This change adds a "-silent" flag to suppress those messages as well as
e metric to count them (so an increase in rate can still be detected).